Add cylinder and plan parsing

+++ srcs/maths/vision_field.c
@@ -6,13 +6,13 @@
 /*   By: mababou <mababou@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/22 15:08:26 by mababou           #+#    #+#             */
-/*   Updated: 2022/04/14 15:07:11 by mababou          ###   ########.fr       */
+/*   Updated: 2022/04/27 11:32:17 by mababou          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/miniRT.h"
 
-int	max_vision_axis(float fov)
+float	max_vision_axis(float fov)
 {
 	float	rad;
